+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Der API Endpoint eines einzelnen Events enthält neu die URL von Anhängen (#1873)
 * Neues "Hitobito Log" einsehbar mit admin permissions unter "Einstellungen" (#1840)
 * Kurse können neu mittels Flag definieren, ob sie die Warteliste verwenden oder nicht (hitobito_cevi#79)
+* Neues Bestätigungsmail für Event Voranmeldung (hitobito_cevi#80)
 
 ## Version 1.28